requirements.txt
alembic==1.7.7
autopep8==1.6.0
click==8.1.3
flask-login==0.6.1
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
flask-wtf==1.0.1
flask==2.1.2
greenlet==1.1.2; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
importlib-metadata==4.11.3; python_version < '3.9'
importlib-resources==5.7.1; python_version < '3.9'
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.0
markupsafe==2.1.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
sqlalchemy==1.4.36
toml==0.10.2
werkzeug==2.1.2
wtforms==3.0.1
zipp==3.8.0; python_version < '3.10'
